refactor: component file name change
